move tests where they belong
